refactor(mvk): move the whole Aeneas-verified surface into nucleus-ifc-kernel (M3 whole-core, Rust)
Completes the kernel carve so EVERY Aeneas proof extraction is single-crate again.
The earlier half-split left CapabilityLattice + decide_pure + the exposure detector
in portcullis-core while their dependencies (Operation/is_exfil/CapabilityLevel)
moved — fracturing the capability + decide-pure single-crate Charon extractions.

Moved into the kernel: capability_lattice.rs (CapabilityLattice + Builder + DimGroup
— the primary Heyting-algebra verification target) and exposure.rs (ExposureLabel/
ExposureSet/classify_operation/project_exposure/should_gate/ifc_to_exposure/
PureVerdict/decide_pure). portcullis-core re-exports them via the existing
`pub use nucleus_ifc_kernel::*`, so its kani harnesses + unit tests (which stay) and
all 15 dependents resolve unchanged. lib.rs is now a thin shell over the kernel +
the unverified machinery.

Fixed broken intra-doc links (crate::capability_traits → plain code spans; that
module stays in portcullis-core) and retargeted doctests/charon comments
portcullis_core → nucleus_ifc_kernel. Ratchet updated (capability_lattice/exposure
enumerated).

Verified: kernel 163 lib + ratchet + 10 doctests; portcullis-core 592; nucleus-ifc
+ nucleus-tool-proxy build; clippy clean. Lean/workflow retarget (capability +
decide-pure extractions → kernel crate) follows next (CI-verified).
