+++ crates/nucleus-ifc-kernel/src/extracted/mod.rs
@@ -14,7 +14,7 @@
 //! tests (see each submodule's `#[cfg(test)]` block).
 //!
 //! The extraction roots live here so the CI extractor can name them with
-//! `charon ... --start-from portcullis_core::extracted::ifc_integrity::<fn>`.
+//! `charon ... --start-from nucleus_ifc_kernel::extracted::ifc_integrity::<fn>`.
 
 pub mod ifc_confidentiality;
 pub mod ifc_integrity;
